test: Add support for expecting master shell commands in fake steps
